requirements.txt
torch>=2.0.0
numpy>=1.21.0
Pillow>=9.0.0
tqdm>=4.65.0  # For progress bars
matplotlib>=3.5.0  # For visualization
pytest>=7.0.0  # For testing 